pyproject.toml
[project]
name = "supportsmith-agent"
version = "0.1.0"
description = "A traceable multi-tool customer support agent with compliance guardrails, self-verification, and multi-turn memory."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "fastapi>=0.115.6",
    "firecrawl-py>=1.6.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langsmith>=0.4.0",
    "openai>=1.59.6",
    "pgvector>=0.3.6",
    "psycopg[binary]>=3.2.3",
    "pydantic-settings>=2.7.1",
    "pyyaml>=6.0.2",
    "sqlalchemy[asyncio]>=2.0.36",
    "uvicorn[standard]>=0.34.0",
]

[project.scripts]
supportsmith-eval = "app.evals.runner:main"
supportsmith-ingest-website = "scripts.ingest_website:main"
supportsmith-migrate = "app.db.migrate:main"
supportsmith-seed = "scripts.db_seed:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "scripts"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.14.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "ruff>=0.9.1",
    "types-PyYAML>=6.0.12.20241230",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = []
explicit_package_bases = true
namespace_packages = true

[[tool.mypy.overrides]]
module = ["pgvector", "pgvector.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
addopts = ["-m", "not live"]
markers = [
    "live: hits the real OpenAI API (and a seeded Postgres). Opt in with `-m live`.",
]
